优化定时器模块

+++ UnityProject/Assets/GameScripts/HotFix/GameLogic/Module/UIModule/UIWindow.cs
@@ -77,7 +77,7 @@ namespace GameLogic
         /// </summary>
         public int HideTimeToClose { get; set; }
         
-        public int HideTimerId { get; set; }
+        public long HideTimerId { get; set; }
 
         public bool IsForever { get; set; }
